+++ 3w_ticketing/src/main/java/com/hhplus/concert_ticketing/infra/user/UserJpaRepository.java
@@ -13,13 +13,13 @@ import java.util.Optional;
 
 public interface UserJpaRepository extends JpaRepository<UserEntity, Long> {
      @Query("SELECT u.balance FROM UserEntity u WHERE u.id = :userId")
-     Optional<Double> findBalanceByUserId(@Param("userId") Long userId);
+     Optional<Integer> findBalanceByUserId(@Param("userId") Long userId);
 
     // 충전 - 낙관적락(기존 구현)
     // @Modifying
     // @Transactional
     // @Query("UPDATE UserEntity u SET u.balance = u.balance + :amount WHERE u.id = :userId AND u.version = :version")
-    // int chargePoint(@Param("userId") Long userId, @Param("amount") double amount, @Param("version") int version);
+    // int chargePoint(@Param("userId") Long userId, @Param("amount") int amount, @Param("version") int version);
 
     // 충전 - 비관적락
     @Query("SELECT u FROM UserEntity u WHERE u.id = :userId")
@@ -29,10 +29,10 @@ public interface UserJpaRepository extends JpaRepository<UserEntity, Long> {
     @Modifying
     @Transactional
     @Query("UPDATE UserEntity u SET u.balance = u.balance + :amount WHERE u.id = :userId")
-    int chargePoint(@Param("userId") Long userId, @Param("amount") double amount);
+    int chargePoint(@Param("userId") Long userId, @Param("amount") int amount);
     
     @Modifying
     @Query("UPDATE UserEntity u SET u.balance = :balance WHERE u.id = :userId AND u.version = :version")
-    int usePoint(@Param("userId") Long userId, @Param("balance") Double balance, @Param("version") int version);
+    int usePoint(@Param("userId") Long userId, @Param("balance") int balance, @Param("version") int version);
 
 }
