+++ 3w_ticketing/src/main/java/com/hhplus/concert_ticketing/presentation/user/ChargeResponse.java
@@ -10,5 +10,5 @@ import lombok.Setter;
 @NoArgsConstructor
 @AllArgsConstructor
 public class ChargeResponse {
-    private Double balance;
+    private int balance;
 }
